add city

+++ src/database/models/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import create_engine, Column, Integer, String, Date, Time
+from sqlalchemy import create_engine, Column, Integer, String, Date, Time, BigInteger
 from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import sessionmaker
 
@@ -8,7 +8,7 @@ Base = declarative_base()
 class UserData(Base):
     __tablename__ = 'user_data'
 
-    id = Column(Integer, primary_key=True, autoincrement=True)
+    id = Column(BigInteger, primary_key=True, autoincrement=True)
     telegram_id = Column(Integer, nullable=False)
     location = Column(String, nullable=False)
     birth_date = Column(Date, nullable=False)
